+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvTimeSeriesMappingSource.java
@@ -8,7 +8,6 @@ package edu.ie3.datamodel.io.source.csv;
 import edu.ie3.datamodel.io.factory.SimpleEntityData;
 import edu.ie3.datamodel.io.factory.timeseries.TimeSeriesMappingFactory;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
-import edu.ie3.datamodel.io.naming.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.source.TimeSeriesMappingSource;
 import java.util.Map;
 import java.util.Optional;
@@ -43,17 +42,14 @@ public class CsvTimeSeriesMappingSource extends CsvDataSource implements TimeSer
     return mapping;
   }
 
-  /** @deprecated since 3.0. Use {@link #timeSeriesMetaInformation(java.util.UUID)} instead */
+  /**
+   * @deprecated since 3.0. Use {@link CsvTimeSeriesTypeSource#getTimeSeriesMetaInformation()}
+   *     instead
+   */
   @Override
   @Deprecated(since = "3.0", forRemoval = true)
   public Optional<edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation>
       getTimeSeriesMetaInformation(UUID timeSeriesUuid) {
     return connector.getIndividualTimeSeriesMetaInformation(timeSeriesUuid);
-  }
-
-  @Override
-  public Optional<IndividualTimeSeriesMetaInformation> timeSeriesMetaInformation(
-      UUID timeSeriesUuid) {
-    return connector.individualTimeSeriesMetaInformation(timeSeriesUuid);
   }
 }