+++ src/main/java/edu/ie3/datamodel/io/source/sql/SqlTimeSeriesSource.java
@@ -9,11 +9,14 @@ import edu.ie3.datamodel.exceptions.SourceException;
 import edu.ie3.datamodel.io.connectors.SqlConnector;
 import edu.ie3.datamodel.io.factory.timeseries.SimpleTimeBasedValueData;
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedSimpleValueFactory;
+import edu.ie3.datamodel.io.naming.DatabaseNamingStrategy;
+import edu.ie3.datamodel.io.naming.timeseries.ColumnScheme;
+import edu.ie3.datamodel.io.naming.timeseries.IndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.io.source.TimeSeriesSource;
-import edu.ie3.datamodel.io.sql.SqlIndividualTimeSeriesMetaInformation;
 import edu.ie3.datamodel.models.timeseries.individual.IndividualTimeSeries;
 import edu.ie3.datamodel.models.timeseries.individual.TimeBasedValue;
 import edu.ie3.datamodel.models.value.Value;
+import edu.ie3.datamodel.utils.TimeSeriesUtils;
 import edu.ie3.util.interval.ClosedInterval;
 import java.sql.Timestamp;
 import java.time.ZonedDateTime;
@@ -22,6 +25,7 @@ import java.util.*;
 public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBasedValue<V>>
     implements TimeSeriesSource<V> {
   private static final String WHERE = " WHERE ";
+  private static final String TIME_SERIES = "time_series";
 
   private final UUID timeSeriesUuid;
   private final Class<V> valueClass;
@@ -41,75 +45,40 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
    *
    * @param connector the connector needed for database connection
    * @param schemaName the database schema to use
-   * @param tableName the database table to use
+   * @param namingStrategy the database entity naming strategy to use
    * @param metaInformation the time series meta information
    * @param timePattern the pattern of time values
    * @return a SqlTimeSeriesSource for given time series table
    * @throws SourceException if the column scheme is not supported
-   * @deprecated since 3.0. Use {@link #getSource(edu.ie3.datamodel.io.connectors.SqlConnector,
-   *     java.lang.String, edu.ie3.datamodel.io.sql.SqlIndividualTimeSeriesMetaInformation,
-   *     java.lang.String)} instead.
    */
-  @Deprecated(since = "3.0", forRemoval = true)
-  public static SqlTimeSeriesSource<? extends Value> getSource(
+  public static SqlTimeSeriesSource<? extends Value> createSource(
       SqlConnector connector,
       String schemaName,
-      String tableName,
-      edu.ie3.datamodel.io.csv.timeseries.IndividualTimeSeriesMetaInformation metaInformation,
+      DatabaseNamingStrategy namingStrategy,
+      IndividualTimeSeriesMetaInformation metaInformation,
       String timePattern)
       throws SourceException {
-    if (!TimeSeriesSource.isSchemeAccepted(metaInformation.getColumnScheme()))
+    if (!TimeSeriesUtils.isSchemeAccepted(metaInformation.getColumnScheme()))
       throw new SourceException(
           "Unsupported column scheme '" + metaInformation.getColumnScheme() + "'.");
 
     Class<? extends Value> valClass = metaInformation.getColumnScheme().getValueClass();
 
     return create(
-        connector, schemaName, tableName, metaInformation.getUuid(), valClass, timePattern);
-  }
-
-  /**
-   * Factory method to build a source from given meta information
-   *
-   * @param connector the connector needed for database connection
-   * @param schemaName the database schema to use
-   * @param metaInformation the time series meta information
-   * @param timePattern the pattern of time values
-   * @return a SqlTimeSeriesSource for given time series table
-   * @throws SourceException if the column scheme is not supported
-   */
-  public static SqlTimeSeriesSource<? extends Value> getSource(
-      SqlConnector connector,
-      String schemaName,
-      SqlIndividualTimeSeriesMetaInformation metaInformation,
-      String timePattern)
-      throws SourceException {
-    if (!TimeSeriesSource.isSchemeAccepted(metaInformation.getColumnScheme()))
-      throw new SourceException(
-          "Unsupported column scheme '" + metaInformation.getColumnScheme() + "'.");
-
-    Class<? extends Value> valClass = metaInformation.getColumnScheme().getValueClass();
-
-    return create(
-        connector,
-        schemaName,
-        metaInformation.getTableName(),
-        metaInformation.getUuid(),
-        valClass,
-        timePattern);
+        connector, schemaName, namingStrategy, metaInformation.getUuid(), valClass, timePattern);
   }
 
   private static <T extends Value> SqlTimeSeriesSource<T> create(
       SqlConnector connector,
       String schemaName,
-      String tableName,
+      DatabaseNamingStrategy namingStrategy,
       UUID timeSeriesUuid,
       Class<T> valClass,
       String timePattern) {
     TimeBasedSimpleValueFactory<T> valueFactory =
         new TimeBasedSimpleValueFactory<>(valClass, timePattern);
     return new SqlTimeSeriesSource<>(
-        connector, schemaName, tableName, timeSeriesUuid, valClass, valueFactory);
+        connector, schemaName, namingStrategy, timeSeriesUuid, valClass, valueFactory);
   }
 
   /**
@@ -117,7 +86,7 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
    *
    * @param connector the connector needed for database connection
    * @param schemaName the database schema to use
-   * @param tableName the database table to use
+   * @param namingStrategy the naming strategy for database entities
    * @param timeSeriesUuid the uuid of the time series
    * @param valueClass the class of returned time series values
    * @param factory a factory that parses the input data
@@ -125,7 +94,7 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
   public SqlTimeSeriesSource(
       SqlConnector connector,
       String schemaName,
-      String tableName,
+      DatabaseNamingStrategy namingStrategy,
       UUID timeSeriesUuid,
       Class<V> valueClass,
       TimeBasedSimpleValueFactory<V> factory) {
@@ -133,13 +102,14 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
     this.timeSeriesUuid = timeSeriesUuid;
     this.valueClass = valueClass;
     this.valueFactory = factory;
+    final ColumnScheme columnScheme = ColumnScheme.parse(valueClass).orElseThrow();
+    final String tableName = namingStrategy.getTimeSeriesEntityName(columnScheme);
 
     String dbTimeColumnName = getDbColumnName(factory.getTimeFieldString(), tableName);
 
-    this.queryFull = createBaseQueryString(schemaName, tableName);
-    this.queryTimeInterval =
-        createQueryStringForTimeInterval(schemaName, tableName, dbTimeColumnName);
-    this.queryTime = createQueryStringForTime(schemaName, tableName, dbTimeColumnName);
+    this.queryFull = createQueryFull(schemaName, tableName);
+    this.queryTimeInterval = createQueryForTimeInterval(schemaName, tableName, dbTimeColumnName);
+    this.queryTime = createQueryForTime(schemaName, tableName, dbTimeColumnName);
   }
 
   @Override
@@ -178,40 +148,69 @@ public class SqlTimeSeriesSource<V extends Value> extends SqlDataSource<TimeBase
    * @return Optional simple time based value
    */
   protected Optional<TimeBasedValue<V>> createEntity(Map<String, String> fieldToValues) {
+    fieldToValues.remove("timeSeries");
     SimpleTimeBasedValueData<V> factoryData =
         new SimpleTimeBasedValueData<>(fieldToValues, valueClass);
     return valueFactory.get(factoryData);
   }
 
   /**
-   * Creates a base query to retrieve all entities in the given time frame with the following
-   * pattern: <br>
-   * {@code <base query> WHERE <time column> BETWEEN ? AND ?;}
+   * Creates a base query to retrieve all entities for this time series: <br>
+   * {@code <base query> WHERE time_series = $timeSeriesUuid AND <time column> BETWEEN ? AND ?;}
+   *
+   * @param schemaName the name of the database schema
+   * @param tableName the name of the database table
+   * @return the query string
+   */
+  private String createQueryFull(String schemaName, String tableName) {
+    return createBaseQueryString(schemaName, tableName)
+        + WHERE
+        + TIME_SERIES
+        + " = '"
+        + timeSeriesUuid.toString()
+        + "'";
+  }
+
+  /**
+   * Creates a base query to retrieve all entities for given time series uuid and in the given time
+   * frame with the following pattern: <br>
+   * {@code <base query> WHERE time_series = $timeSeriesUuid AND <time column> BETWEEN ? AND ?;}
    *
    * @param schemaName the name of the database schema
    * @param tableName the name of the database table
    * @param timeColumnName the name of the column holding the timestamp info
    * @return the query string
    */
-  private static String createQueryStringForTimeInterval(
+  private String createQueryForTimeInterval(
       String schemaName, String tableName, String timeColumnName) {
     return createBaseQueryString(schemaName, tableName)
         + WHERE
+        + TIME_SERIES
+        + " = '"
+        + timeSeriesUuid.toString()
+        + "' AND "
         + timeColumnName
         + " BETWEEN ? AND ?;";
   }
 
   /**
-   * Creates a basic query to retrieve an entry for the given time with the following pattern: <br>
-   * {@code <base query> WHERE <time column>=?;}
+   * Creates a basic query to retrieve an entry for the given time series uuid and time with the
+   * following pattern: <br>
+   * {@code <base query> WHERE time_series = $timeSeriesUuid AND <time column>=?;}
    *
    * @param schemaName the name of the database schema
-   * @param weatherTableName the name of the database table
+   * @param tableName the name of the database table
    * @param timeColumnName the name of the column holding the timestamp info
    * @return the query string
    */
-  private String createQueryStringForTime(
-      String schemaName, String weatherTableName, String timeColumnName) {
-    return createBaseQueryString(schemaName, weatherTableName) + WHERE + timeColumnName + "=?;";
+  private String createQueryForTime(String schemaName, String tableName, String timeColumnName) {
+    return createBaseQueryString(schemaName, tableName)
+        + WHERE
+        + TIME_SERIES
+        + " = '"
+        + timeSeriesUuid.toString()
+        + "' AND "
+        + timeColumnName
+        + "=?;";
   }
 }